requirements.txt
streamlit==1.37.1
pandas==2.2.2
numpy==1.26.4
plotly==5.23.0
ccxt==4.3.98
ta==0.11.0
pydantic==2.8.2
cryptography==43.0.1
SQLAlchemy==2.0.32
python-dotenv==1.0.1
streamlit-authenticator==0.4.1
streamlit-oauth==0.1.14
requests==2.32.3
scipy==1.13.1
